Wire all frontend modules to backend; fix validation and DB constraints
Backend:
- Add /api/health endpoint
- Insurance: add 'life','ulip' types; add 'yearly' premium_frequency
- Credit card: accept date string for bill_due_date (extract day-of-month)
- Rental: add residential/pg/land/custom property_type values
- Fix DB CHECK constraints for insurance type, frequency, rental property_type
- Fix PostgreSQL type-inference error in loan and lending UPDATE queries
  (add ::numeric cast to $1 used in CASE expressions)

Frontend (fincare-web.html):
- loadDashboardData: sync State.loans, State.ccCards, State.rentalProperties
  from API data with field-name mapping so detail screens render correctly
- Map lendBorrow API fields (person_name→person, notes→reason) for renderLendBorrow
- Wire _loanSave() → POST /api/loans
- Wire _loanSaveTx() → POST /api/loans/:id/transactions
- Wire _ccSaveTransaction() → POST /api/credit-cards/:id/transactions
- Wire _rentalSaveProperty() → POST /api/rentals
- Wire _rentalRecordRent() → POST /api/rentals/:id/transactions
- Add _lbShowAddForm(), _lbSave(), _lbRecordPayment(), _lbDoPayment()
  for lending/borrowing create and payment with modal UI
- renderLendBorrow: add + Add button, empty-state message, Record Payment button

+++ backend/src/controllers/lending.controller.js
@@ -99,8 +99,8 @@ const recordPayment = asyncWrapper(async (req, res) => {
 
     const newRemaining = Math.max(0, parseFloat(entryRes.rows[0].remaining) - parseFloat(amount));
     await client.query(
-      `UPDATE lending SET remaining = $1,
-        status = CASE WHEN $1 <= 0 THEN 'settled' ELSE CASE WHEN $1 < amount THEN 'partial' ELSE status END END
+      `UPDATE lending SET remaining = $1::numeric,
+        status = CASE WHEN $1::numeric <= 0 THEN 'settled' ELSE CASE WHEN $1::numeric < amount THEN 'partial' ELSE status END END
        WHERE id = $2`,
       [newRemaining, req.params.id]
     );
